refactor(all): refactor to split to setUp, eventLoop and shutDown 3 phases
