+++ Server.hpp
@@ -1,8 +1,6 @@
 #ifndef SERVER_HPP
 #define SERVER_HPP
 
-#define READ_BUFF_SIZE 1024
-#define POLL_TIMEOUT 1000
 
 #include <cerrno>
 #include <cstdlib>
@@ -16,12 +14,6 @@
 #include "Client.hpp"
 #include <netinet/in.h>
 
-typedef struct Data
-{
-	std::vector<Server*> servers;
-	std::vector<Client*> clients;
-	std::vector<pollfd> fdPool;
-}	Data;
 
 class Server
 {
@@ -52,7 +44,4 @@ class Server
 std::ostream& operator<<(std::ostream& out, Server& server);
 Server* findServer(int fd, std::vector<Server*>& servers);
 pollfd fdToPollfdWithStatus(int fd, short event);
-void	shutDownServers(std::vector<Server*>& servers);
-void setUpServers(Data& data);
-void pollEventsLoop(Data& data);
 #endif // !SERVER_HPP
